+++ src/linksmart.eu/lc/core/catalog/resource/catalogapi.go
@@ -116,7 +116,7 @@ func (self *Resource) unLdify(apiLocation string) Resource {
 
 func (self ReadableCatalogAPI) collectionFromDevices(devices []Device, page, perPage, total int) *Collection {
 	respDevices := make(map[string]EmptyDevice)
-	respResources := make([]Resource, 0, self.catalogStorage.getResourcesCount())
+	var respResources []Resource
 
 	for _, d := range devices {
 		dld := d.ldify(self.apiLocation)
@@ -170,16 +170,50 @@ func (self ReadableCatalogAPI) paginatedDeviceFromDevice(d Device, page, perPage
 	return pd
 }
 
+// Error describes an API error (serializable in JSON)
+type Error struct {
+	// Code is the (http) code of the error
+	Code int `json:"code"`
+	// Message is the (human-readable) error message
+	Message string `json:"message"`
+}
+
+// ErrorResponse writes error to HTTP ResponseWriter
+func ErrorResponse(w http.ResponseWriter, code int, msgs ...string) {
+	msg := strings.Join(msgs, " ")
+	e := &Error{
+		code,
+		msg,
+	}
+	logger.Println("ERROR:", msg)
+	b, _ := json.Marshal(e)
+	w.Header().Set("Content-Type", "application/json;version="+ApiVersion)
+	w.WriteHeader(code)
+	w.Write(b)
+}
+
 func (self ReadableCatalogAPI) List(w http.ResponseWriter, req *http.Request) {
-	req.ParseForm()
+	err := req.ParseForm()
+	if err != nil {
+		ErrorResponse(w, http.StatusBadRequest, "Error parsing the query:", err.Error())
+		return
+	}
 	page, _ := strconv.Atoi(req.Form.Get(GetParamPage))
 	perPage, _ := strconv.Atoi(req.Form.Get(GetParamPerPage))
 	page, perPage = catalog.ValidatePagingParams(page, perPage, MaxPerPage)
 
-	devices, total, _ := self.catalogStorage.getMany(page, perPage)
+	devices, total, err := self.catalogStorage.getMany(page, perPage)
+	if err != nil {
+		ErrorResponse(w, http.StatusInternalServerError, err.Error())
+		return
+	}
 	coll := self.collectionFromDevices(devices, page, perPage, total)
 
-	b, _ := json.Marshal(coll)
+	b, err := json.Marshal(coll)
+	if err != nil {
+		ErrorResponse(w, http.StatusInternalServerError, err.Error())
+		return
+	}
 	w.Header().Set("Content-Type", "application/ld+json;version="+ApiVersion)
 	w.Write(b)
 
@@ -193,14 +227,17 @@ func (self ReadableCatalogAPI) Filter(w http.ResponseWriter, req *http.Request) 
 	fop := params["op"]
 	fvalue := params["value"]
 
-	req.ParseForm()
+	err := req.ParseForm()
+	if err != nil {
+		ErrorResponse(w, http.StatusBadRequest, "Error parsing the query:", err.Error())
+		return
+	}
 	page, _ := strconv.Atoi(req.Form.Get(GetParamPage))
 	perPage, _ := strconv.Atoi(req.Form.Get(GetParamPerPage))
 	page, perPage = catalog.ValidatePagingParams(page, perPage, MaxPerPage)
 
 	var (
 		data  interface{}
-		err   error
 		total int
 	)
 
@@ -231,32 +268,37 @@ func (self ReadableCatalogAPI) Filter(w http.ResponseWriter, req *http.Request) 
 
 	case FTypeResources:
 		data, total, err = self.catalogStorage.pathFilterResources(fpath, fop, fvalue, page, perPage)
-		devs := self.catalogStorage.devicesFromResources(data.([]Resource))
-		data = self.collectionFromDevices(devs, page, perPage, total)
+		data = self.collectionFromDevices(data.([]Device), page, perPage, total)
 		if data.(*Collection).Total == 0 {
 			data = nil
 		}
 	}
 
 	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprintf(w, "Error processing the request: %s\n", err.Error())
+		ErrorResponse(w, http.StatusInternalServerError, "Error processing the request:", err.Error())
 		return
 	}
 
 	if data == nil {
-		w.WriteHeader(http.StatusNotFound)
-		fmt.Fprintf(w, "No matched entries found\n")
+		ErrorResponse(w, http.StatusNotFound, "No matched entries found.")
 		return
 	}
 
-	b, _ := json.Marshal(data)
+	b, err := json.Marshal(data)
+	if err != nil {
+		ErrorResponse(w, http.StatusInternalServerError, err.Error())
+		return
+	}
 	w.Header().Set("Content-Type", "application/ld+json;version="+ApiVersion)
 	w.Write(b)
 }
 
 func (self ReadableCatalogAPI) Get(w http.ResponseWriter, req *http.Request) {
-	req.ParseForm()
+	err := req.ParseForm()
+	if err != nil {
+		ErrorResponse(w, http.StatusBadRequest, "Error parsing the query:", err.Error())
+		return
+	}
 	page, _ := strconv.Atoi(req.Form.Get(GetParamPage))
 	perPage, _ := strconv.Atoi(req.Form.Get(GetParamPerPage))
 	page, perPage = catalog.ValidatePagingParams(page, perPage, MaxPerPage)
@@ -266,17 +308,19 @@ func (self ReadableCatalogAPI) Get(w http.ResponseWriter, req *http.Request) {
 
 	d, err := self.catalogStorage.get(id)
 	if err == ErrorNotFound {
-		w.WriteHeader(http.StatusNotFound)
-		fmt.Fprintf(w, "Device not found\n")
+		ErrorResponse(w, http.StatusNotFound, "Device not found.")
 		return
 	} else if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprintf(w, "Error requesting the device: %s\n", err.Error())
+		ErrorResponse(w, http.StatusInternalServerError, "Error requesting the device:", err.Error())
 		return
 	}
 
 	pd := self.paginatedDeviceFromDevice(d, page, perPage)
-	b, _ := json.Marshal(pd)
+	b, err := json.Marshal(pd)
+	if err != nil {
+		ErrorResponse(w, http.StatusInternalServerError, err.Error())
+		return
+	}
 
 	w.Header().Set("Content-Type", "application/ld+json;version="+ApiVersion)
 	w.Write(b)
@@ -291,28 +335,29 @@ func (self ReadableCatalogAPI) GetResource(w http.ResponseWriter, req *http.Requ
 	// check if device devid exists
 	_, err := self.catalogStorage.get(devid)
 	if err == ErrorNotFound {
-		w.WriteHeader(http.StatusNotFound)
-		fmt.Fprintf(w, "Registration not found\n")
+		ErrorResponse(w, http.StatusNotFound, "Registration not found.")
 		return
 	} else if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprintf(w, "Error requesting the device: %s\n", err.Error())
+		ErrorResponse(w, http.StatusInternalServerError, "Error requesting the device:", err.Error())
 		return
 	}
 
 	// check if it has a resource resid
 	res, err := self.catalogStorage.getResourceById(resid)
 	if err == ErrorNotFound {
-		w.WriteHeader(http.StatusNotFound)
-		fmt.Fprintf(w, "Registration not found\n")
+		ErrorResponse(w, http.StatusNotFound, "Registration not found.")
 		return
 	} else if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprintf(w, "Error requesting the resource: %s\n", err.Error())
+		ErrorResponse(w, http.StatusInternalServerError, "Error requesting the resource:", err.Error())
 		return
 	}
 
-	b, _ := json.Marshal(res.ldify(self.apiLocation))
+	b, err := json.Marshal(res.ldify(self.apiLocation))
+	if err != nil {
+		ErrorResponse(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+
 	w.Header().Set("Content-Type", "application/ld+json;version="+ApiVersion)
 	w.Write(b)
 	return
@@ -320,20 +365,22 @@ func (self ReadableCatalogAPI) GetResource(w http.ResponseWriter, req *http.Requ
 
 func (self WritableCatalogAPI) Add(w http.ResponseWriter, req *http.Request) {
 	body, err := ioutil.ReadAll(req.Body)
+	if err != nil {
+		ErrorResponse(w, http.StatusBadRequest, err.Error())
+		return
+	}
 	req.Body.Close()
 
 	var d Device
 	err = json.Unmarshal(body, &d)
 	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprintf(w, "Error processing the request: %s\n", err.Error())
+		ErrorResponse(w, http.StatusBadRequest, "Error processing the request:", err.Error())
 		return
 	}
 
 	err = self.catalogStorage.add(d)
 	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprintf(w, "Error creating the registration: %s\n", err.Error())
+		ErrorResponse(w, http.StatusInternalServerError, "Error creating the registration:", err.Error())
 		return
 	}
 
@@ -348,24 +395,25 @@ func (self WritableCatalogAPI) Update(w http.ResponseWriter, req *http.Request) 
 	id := fmt.Sprintf("%v/%v", params["dgwid"], params["regid"])
 
 	body, err := ioutil.ReadAll(req.Body)
+	if err != nil {
+		ErrorResponse(w, http.StatusBadRequest, err.Error())
+		return
+	}
 	req.Body.Close()
 
 	var d Device
 	err = json.Unmarshal(body, &d)
 	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprintf(w, "Error processing the request:: %s\n", err.Error())
+		ErrorResponse(w, http.StatusBadRequest, "Error processing the request:", err.Error())
 		return
 	}
 
 	err = self.catalogStorage.update(id, d)
 	if err == ErrorNotFound {
-		w.WriteHeader(http.StatusNotFound)
-		fmt.Fprintf(w, "Not found\n")
+		ErrorResponse(w, http.StatusNotFound, "Not found.")
 		return
 	} else if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprintf(w, "Error updating the device: %s\n", err.Error())
+		ErrorResponse(w, http.StatusInternalServerError, "Error updating the device:", err.Error())
 		return
 	}
 
@@ -380,12 +428,10 @@ func (self WritableCatalogAPI) Delete(w http.ResponseWriter, req *http.Request) 
 
 	err := self.catalogStorage.delete(id)
 	if err == ErrorNotFound {
-		w.WriteHeader(http.StatusNotFound)
-		fmt.Fprintf(w, "Not found\n")
+		ErrorResponse(w, http.StatusNotFound, "Not found.")
 		return
 	} else if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprintf(w, "Error deleting the device: %s\n", err.Error())
+		ErrorResponse(w, http.StatusInternalServerError, "Error deleting the device:", err.Error())
 		return
 	}