+++ src/linksmart.eu/lc/core/catalog/service/catalog.go
@@ -22,7 +22,7 @@ type Service struct {
 	Ttl            int                    `json:"ttl"`
 	Created        time.Time              `json:"created"`
 	Updated        time.Time              `json:"updated"`
-	Expires        time.Time              `json:"expires"`
+	Expires        *time.Time             `json:"expires"`
 }
 
 // Deep copy of the Service
@@ -86,8 +86,9 @@ type CatalogStorage interface {
 
 	// Utility functions
 	getMany(page, perPage int) ([]Service, int, error)
-	getCount() int
+	getCount() (int, error)
 	cleanExpired(ts time.Time)
+	Close() error
 
 	// Path filtering
 	pathFilterOne(path, op, value string) (Service, error)