+++ src/linksmart.eu/lc/core/cmd/device-gateway/config.go
@@ -11,6 +11,7 @@ import (
 	"strings"
 	"time"
 
+	utils "linksmart.eu/lc/core/catalog"
 	"linksmart.eu/lc/sec/authz"
 )
 
@@ -105,6 +106,7 @@ type Config struct {
 	StaticDir      string                       `json:"staticDir`
 	Catalog        []Catalog                    `json:"catalog"`
 	Http           HttpConfig                   `json:"http"`
+	Storage        StorageConfig                `json:"storage"`
 	Protocols      map[ProtocolType]interface{} `json:"protocols"`
 	Devices        []Device                     `json:"devices"`
 	Auth           ValidatorConf                `json:"auth"`
@@ -123,6 +125,12 @@ func (c *Config) Validate() error {
 
 	// Check if HTTP configuration is valid
 	err = c.Http.Validate()
+	if err != nil {
+		return err
+	}
+
+	// Check if Storage configuration is valid
+	err = c.Storage.Validate()
 	if err != nil {
 		return err
 	}
@@ -212,6 +220,25 @@ type HttpConfig struct {
 func (h *HttpConfig) Validate() error {
 	if h.BindAddr == "" || h.BindPort == 0 {
 		return fmt.Errorf("HTTP bindAddr and bindPort have to be defined")
+	}
+	return nil
+}
+
+//
+// Storage config
+//
+type StorageConfig struct {
+	Type string `json:"type"`
+}
+
+var supportedBackends = map[string]bool{
+	utils.CatalogBackendMemory:  true,
+	utils.CatalogBackendLevelDB: true,
+}
+
+func (c *StorageConfig) Validate() error {
+	if !supportedBackends[c.Type] {
+		return fmt.Errorf("Unsupported storage backend")
 	}
 	return nil
 }